fix sonarqube issue

+++ services/ble/Gap.cpp
@@ -7,7 +7,7 @@ namespace
 {
     void AddHeader(infra::BoundedVector<uint8_t>& payload, std::size_t length, services::GapAdvertisementDataType type)
     {
-        payload.push_back(length + 1);
+        payload.push_back(static_cast<uint8_t>(length + 1));
         payload.push_back(static_cast<uint8_t>(type));
     }
